requirements.txt
Flask
Flast-WTF
Flask-SQLAlchemy
Flask-login
gunicorn